feat: move bridges into layout folder
