+++ src/ui/layouts/ViewerPageLayout.tsx
@@ -1,7 +1,7 @@
 import { Outlet } from "react-router-dom";
 import { Footer } from "../components/common/footer/Footer";
 import { Header } from "./header/Header";
-import { ViewerPageBridge } from "../providers/ViewerPageBridge";
+import { ViewerPageBridge } from "./menu-bridges/ViewerPageBridge";
 import { ServerStatus } from "../features/system/components/server-status/ServerStatus";
 
 import "./Layout.css";
